catalogo-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.8.2
pydantic-settings>=2.2,<3
SQLAlchemy[asyncio]==2.0.35
redis==5.1.1
orjson==3.10.7
python-dotenv==1.0.1
asyncpg>=0.29,<1
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
pytest
pytest-cov
pytest-asyncio
httpx
fakeredis
freezegun
coverage
aiosqlite
